reorganize files, adjust pathing to scss, create palette
