+++ client/src/ASSETS/avatar.jpg
