+++ client/src/SASS/Login.scss
@@ -1,6 +1,6 @@
 .Login-container {
   align-items: center;
-  background: url('./ASSETS/loginpage-bg.jpg');
+  background: url('../ASSETS/loginpage-bg.jpg');
   color: white;
   display: flex;
   height: 100vh;
