+++ client/src/SASS/_palette.scss
@@ -1,0 +1,14 @@
+// UX PALETTE
+$white: #ffffff;
+$black: #15161a;
+$dark-grey: #6d6d72;
+$blue: #5557fe;
+$purple: #9555fe;
+$btn-gradient: linear-gradient(99.98deg, #5557fe -2.36%, #9555fe 103.67%);
+
+// LIGHT MODE
+$light-grey: #f4f4f4;
+$grey: #d8d8d8;
+
+//DARK MODE
+$almost-black: #212229;
