+++ client/src/components/UserProfile_LI.js
@@ -3,11 +3,11 @@ import { Link } from 'react-router-dom';
 
 import Heatmap from './Heatmap.js';
 
-import './SASS/UserProfile.scss';
+import '../SASS/UserProfile.scss';
 
-import avatar from '../components/ASSETS/avatar.jpg';
-// import link from '../components/ASSETS/link.svg';
-// import map from '../components/ASSETS/map-pin.svg';
+import avatar from '../ASSETS/avatar.jpg';
+// import link from '../ASSETS/link.svg';
+// import map from '../ASSETS/map-pin.svg';
 import UserProfileTabs from './UserProfile_Tabs';
 
 class UserProfile_LI extends Component {
